django_backend/requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.5.0
djangorestframework==3.11.0
et-xmlfile==1.0.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
xlrd==2.0.1
XlsxWriter==1.4.0
